Add files via upload
This workflow is like magic. Because by using AI + n8n, we can use AI's ability to summarize what happened in the emails. We chat on Telegram and the feedback is coming from Telegram also.

Here is the explanation :
From the telegram input, we are using an AI agent for  Natural Language Processing to translate the telegram input as an SQL query. After that, the generated query from the AI agent will be processed by PostgreSQL as an executed query. Next, we create another AI agent to generate some information from an SQL query result, such as :

1. Group similar messages and show the count.
2. Summarize them clearly.
3. Highlight any error messages that appear to be new or unusual based on tone or phrasing.